[RENAME/#111] 패키지명 수정
