+++ app/src/main/java/com/napzak/market/data/product/di/RepositoryModule.kt
@@ -2,8 +2,8 @@ package com.napzak.market.data.product.di
 
 import com.napzak.market.data.product.repositoryimpl.ExploreProductRepositoryImpl
 import com.napzak.market.data.product.repositoryimpl.MarketProductRepositoryImpl
-import com.napzak.market.domain.explore.repository.ExploreProductRepository
-import com.napzak.market.domain.explore.repository.MarketProductRepository
+import com.napzak.market.domain.product.repository.ExploreProductRepository
+import com.napzak.market.domain.product.repository.MarketProductRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
