+++ app/src/main/java/com/napzak/market/data/product/dto/request/MarketProductBuyItemsRequest.kt
@@ -1,10 +1,10 @@
-package com.napzak.market.data.marketinfo.dto
+package com.napzak.market.data.product.dto.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class MarketProductSellItemsRequest(
+data class MarketProductBuyItemsRequest(
     @SerialName("storeOwnerId")
     val storeId: Long?,
     @SerialName("sortOption")
@@ -13,6 +13,4 @@ data class MarketProductSellItemsRequest(
     val genreIds: List<Long>?,
     @SerialName("isOnSale")
     val isOnSale: Boolean,
-    @SerialName("isUnopened")
-    val isUnopened: Boolean,
 )
