+++ app/src/main/java/com/napzak/market/data/product/dto/request/MarketProductSellItemsRequest.kt
@@ -1,10 +1,10 @@
-package com.napzak.market.data.marketinfo.dto
+package com.napzak.market.data.product.dto.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class MarketProductBuyItemsRequest(
+data class MarketProductSellItemsRequest(
     @SerialName("storeOwnerId")
     val storeId: Long?,
     @SerialName("sortOption")
@@ -13,4 +13,6 @@ data class MarketProductBuyItemsRequest(
     val genreIds: List<Long>?,
     @SerialName("isOnSale")
     val isOnSale: Boolean,
+    @SerialName("isUnopened")
+    val isUnopened: Boolean,
 )
