+++ app/src/main/java/com/napzak/market/data/product/mapper/ExploreSearchedBuyItemsMapper.kt
@@ -1,7 +1,7 @@
 package com.napzak.market.data.product.mapper
 
 import com.napzak.market.data.product.dto.request.SearchedProductBuyItemsRequest
-import com.napzak.market.domain.explore.model.ProductListFilterWithSearchTerm
+import com.napzak.market.domain.product.model.ProductListFilterWithSearchTerm
 
 fun ProductListFilterWithSearchTerm.toSearchedProductBuyRequest(): SearchedProductBuyItemsRequest =
     SearchedProductBuyItemsRequest(
