+++ app/src/main/java/com/napzak/market/data/product/mapper/ExploreSearchedSellItemsMapper.kt
@@ -1,7 +1,7 @@
 package com.napzak.market.data.product.mapper
 
 import com.napzak.market.data.product.dto.request.SearchedProductSellItemsRequest
-import com.napzak.market.domain.explore.model.ProductListFilterWithSearchTerm
+import com.napzak.market.domain.product.model.ProductListFilterWithSearchTerm
 
 fun ProductListFilterWithSearchTerm.toSearchedProductSellRequest(): SearchedProductSellItemsRequest =
     SearchedProductSellItemsRequest(
