+++ app/src/main/java/com/napzak/market/data/product/mapper/MarketProductBuyItemsMapper.kt
@@ -1,7 +1,7 @@
 package com.napzak.market.data.product.mapper
 
 import com.napzak.market.data.product.dto.request.MarketProductBuyItemsRequest
-import com.napzak.market.domain.explore.model.ProductListFilter
+import com.napzak.market.domain.product.model.ProductListFilter
 
 fun ProductListFilter.toMarketProductBuyRequest(): MarketProductBuyItemsRequest =
     MarketProductBuyItemsRequest(
