+++ app/src/main/java/com/napzak/market/data/product/repositoryimpl/MarketProductRepositoryImpl.kt
@@ -1,29 +1,28 @@
 package com.napzak.market.data.product.repositoryimpl
 
-import com.napzak.market.data.marketinfo.mapper.toMarketProductBuyItems
-import com.napzak.market.data.marketinfo.mapper.toMarketProductBuyRequest
-import com.napzak.market.data.marketinfo.mapper.toMarketProductSellItems
-import com.napzak.market.data.marketinfo.mapper.toMarketProductSellRequest
+import com.napzak.market.data.product.mapper.toMarketProductBuyRequest
+import com.napzak.market.data.product.mapper.toMarketProductSellRequest
 import com.napzak.market.data.product.datasource.MarketProductDataSource
-import com.napzak.market.domain.explore.repository.MarketProductRepository
-import com.napzak.market.domain.explore.model.ProductListFilter
-import com.napzak.market.domain.explore.model.ProductItem
+import com.napzak.market.data.product.mapper.toProductItems
+import com.napzak.market.domain.product.repository.MarketProductRepository
+import com.napzak.market.domain.product.model.ProductListFilter
+import com.napzak.market.domain.product.model.Product
 import javax.inject.Inject
 
 class MarketProductRepositoryImpl @Inject constructor(
     private val marketProductDataSource: MarketProductDataSource,
 ) : MarketProductRepository {
-    override suspend fun fetchMarketProductSellItems(request: ProductListFilter): Result<List<ProductItem>> =
+    override suspend fun fetchMarketProductSellItems(request: ProductListFilter): Result<List<Product>> =
         runCatching {
             val response =
                 marketProductDataSource.getMarketProductSellItems(request.toMarketProductSellRequest())
-            response.data.toMarketProductSellItems()
+            response.data.productSellList.toProductItems()
         }
 
-    override suspend fun fetchMarketProductBuyItems(request: ProductListFilter): Result<List<ProductItem>> =
+    override suspend fun fetchMarketProductBuyItems(request: ProductListFilter): Result<List<Product>> =
         runCatching {
             val response =
                 marketProductDataSource.getMarketProductBuyItems(request.toMarketProductBuyRequest())
-            response.data.toMarketProductBuyItems()
+            response.data.productBuyList.toProductItems()
         }
 }
