+++ app/src/main/java/com/napzak/market/domain/product/model/ProductListFilter.kt
@@ -1,4 +1,4 @@
-package com.napzak.market.domain.explore.model
+package com.napzak.market.domain.product.model
 
 data class ProductListFilter(
     val storeId: Long? = null,
