+++ app/src/main/java/com/napzak/market/domain/product/repository/ExploreProductRepository.kt
@@ -1,8 +1,8 @@
-package com.napzak.market.domain.explore.repository
+package com.napzak.market.domain.product.repository
 
-import com.napzak.market.domain.explore.model.Product
-import com.napzak.market.domain.explore.model.ProductListFilter
-import com.napzak.market.domain.explore.model.ProductListFilterWithSearchTerm
+import com.napzak.market.domain.product.model.Product
+import com.napzak.market.domain.product.model.ProductListFilter
+import com.napzak.market.domain.product.model.ProductListFilterWithSearchTerm
 
 interface ExploreProductRepository {
     suspend fun fetchSellProductItems(request: ProductListFilter): Result<List<Product>>
