+++ app/src/main/java/com/napzak/market/domain/product/repository/MarketProductRepository.kt
@@ -1,7 +1,7 @@
-package com.napzak.market.domain.explore.repository
+package com.napzak.market.domain.product.repository
 
-import com.napzak.market.domain.explore.model.Product
-import com.napzak.market.domain.explore.model.ProductListFilter
+import com.napzak.market.domain.product.model.Product
+import com.napzak.market.domain.product.model.ProductListFilter
 
 interface MarketProductRepository {
     suspend fun fetchMarketProductSellItems(request: ProductListFilter): Result<List<Product>>
