+++ app/src/main/java/com/napzak/market/domain/product/usecase/MarketProductBuyItemsUseCase.kt
@@ -1,34 +1,33 @@
-package com.napzak.market.domain.marketinfo.usecase
+package com.napzak.market.domain.product.usecase
 
+import com.napzak.market.domain.product.model.Product
+import com.napzak.market.domain.product.model.ProductListFilter
+import com.napzak.market.domain.product.repository.MarketProductRepository
 import com.napzak.market.domain.genre.model.Genre
-import com.napzak.market.domain.marketinfo.model.ProductItem
-import com.napzak.market.domain.marketinfo.model.ProductListFilter
-import com.napzak.market.domain.marketinfo.repository.MarketInfoRepository
 import javax.inject.Inject
 
-class MarketProductSellItemsUseCase @Inject constructor(
-    private val marketInfoRepository: MarketInfoRepository,
+class MarketProductBuyItemsUseCase @Inject constructor(
+    private val marketProductRepository: MarketProductRepository,
 ) {
     suspend operator fun invoke(
         storeId: Long,
         sortType: String,
         genreItems: List<Genre>?,
         isOnSale: Boolean,
-        isUnopened: Boolean,
-    ): Result<List<ProductItem>> {
+    ): Result<List<Product>> {
         val genreIds = if (genreItems?.isNotEmpty() == true) {
             genreItems.map { it.genreId }
         } else {
             null
         }
 
-        return marketInfoRepository.fetchMarketProductSellItems(
+        return marketProductRepository.fetchMarketProductBuyItems(
             ProductListFilter(
                 storeId = storeId,
                 sortOption = sortType,
                 genreId = genreIds,
                 isOnSale = isOnSale,
-                isUnopened = isUnopened,
+                isUnopened = false,
             )
         )
     }
