+++ app/src/main/java/com/napzak/market/presentation/marketinfo/MarketInfoViewModel.kt
@@ -10,8 +10,8 @@ import com.napzak.market.domain.genre.model.Genre
 import com.napzak.market.domain.genre.usecase.GenreSearchUseCase
 import com.napzak.market.domain.interest.repository.InterestRepository
 import com.napzak.market.domain.marketinfo.repository.MarketInfoRepository
-import com.napzak.market.domain.marketinfo.usecase.MarketProductBuyItemsUseCase
-import com.napzak.market.domain.marketinfo.usecase.MarketProductSellItemsUseCase
+import com.napzak.market.domain.product.usecase.MarketProductBuyItemsUseCase
+import com.napzak.market.domain.product.usecase.MarketProductSellItemsUseCase
 import com.napzak.market.presentation.marketinfo.state.MarketInfoBottomSheetState
 import com.napzak.market.presentation.marketinfo.state.MarketInfoUiState
 import com.napzak.market.presentation.marketinfo.state.MarketProductItemsInformation
